// front-end/webapp/src/app/Interfaces/ListParameterResponse.ts
export interface ListParametersResponse {
    spaceSystems: string[];
    parameters: ParameterInfo[];
  
    // Token indicating the response is only partial. More results can then
    // be obtained by performing the same request (including all original
    // query parameters) and setting the ``next`` parameter to this token.
    continuationToken: string;
  
    // The total number of results (across all pages)
    totalSize: number;
}

export interface ParameterInfo {
    name: string;
    qualifiedName: string;
    shortDescription: string;
    longDescription: string;
    alias: NamedObjectId[];
    type: ParameterTypeInfo;
    dataSource: DataSourceType;
    usedBy: UsedByInfo;
    ancillaryData: {[key: string]: any};
  
    // Operations that return aggregate members or array entries
    // may use this field to indicate the path within the parameter.
    path: string[];
}

// Used by external clients to identify an item in the Mission Database
// If namespace is set, then the name is that of an alias, rather than
// the qualified name.
export interface NamedObjectId {
    name: string;
    namespace: string;
}


export interface ParameterTypeInfo {
    engType: string;
    dataEncoding: DataEncodingInfo;
    unitSet: UnitInfo[];
    defaultAlarm: AlarmInfo;
    enumValue: EnumValue[];
    absoluteTimeInfo: AbsoluteTimeInfo;
    contextAlarm: ContextAlarmInfo[];
    member: MemberInfo[];
    arrayInfo: ArrayInfo;
    ancillaryData: {[key: string]: any};
}
// ...
export interface UsedByInfo {
    algorithm: AlgorithmInfo[];
    container: ContainerInfo[];
}
// ...
export enum DataSourceType {
    TELEMETERED = "TELEMETERED",
    DERIVED = "DERIVED",
    CONSTANT = "CONSTANT",
    LOCAL = "LOCAL",
    SYSTEM = "SYSTEM",
    COMMAND = "COMMAND",
    COMMAND_HISTORY = "COMMAND_HISTORY",
    EXTERNAL1 = "EXTERNAL1",
    EXTERNAL2 = "EXTERNAL2",
    EXTERNAL3 = "EXTERNAL3",
}